LayoutTests/platform/gtk/http/tests/media/video-auth-expected.txt

didReceiveAuthenticationChallenge - Responding with username:password
 
Tests that the media player sends authorization credentials when requesting a media file.  
EVENT(canplay)
END OF TEST
